Increase minSdkVersion to 14 and merge implementations
AndroidX needs at least API level 14 (Ice Cream Sandwich) which was
released 10 years ago and should not exclude many devices. Having a
merged tree simplifies the development.
