+++ brouter-routing-app/build.gradle
@@ -13,6 +13,8 @@ android {
 
         resValue('string', 'app_version', defaultConfig.versionName)
         setProperty("archivesBaseName","BRouterApp." + defaultConfig.versionName)
+
+        minSdkVersion 14
     }
 
 	if(project.hasProperty("RELEASE_STORE_FILE")) {
@@ -70,13 +72,11 @@ android {
         api19 {
             dimension "api"
 
-            minSdkVersion 10
             targetSdkVersion 19
         }
         api30 {
             dimension "api"
 
-            minSdkVersion 19
             targetSdkVersion 30
         }
     }
@@ -85,7 +85,7 @@ android {
 
 dependencies {
 
-    api30Implementation 'androidx.appcompat:appcompat:1.3.1'
+    implementation 'androidx.appcompat:appcompat:1.3.1'
 
     implementation project(':brouter-mapaccess')
     implementation project(':brouter-core')
